Add sample for blob_client
