--- sample/Makefile
@@ -1,3 +1,0 @@
-
-all:
-	g++ --std=c++11 sample.cpp -o run -I../include -L.. -lazure-storage -lcurl
